+++ source/IntegrationTests/Infrastructure/InboxEvents/TestInboxEventMapper.cs
@@ -16,7 +16,7 @@ using System;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
-using Energinet.DataHub.EDI.Infrastructure.InboxEvents;
+using Energinet.DataHub.EDI.Process.Infrastructure.InboxEvents;
 using MediatR;
 
 namespace Energinet.DataHub.EDI.IntegrationTests.Infrastructure.InboxEvents;
