+++ source/IntegrationTests/Infrastructure/InboxEvents/WhenAggregatedTimeSeriesRequestAcceptedEventIsReceivedTests.cs
@@ -18,11 +18,11 @@ using System.Threading.Tasks;
 using Dapper;
 using Energinet.DataHub.EDI.BuildingBlocks.Infrastructure.DataAccess;
 using Energinet.DataHub.EDI.Common.DateTime;
-using Energinet.DataHub.EDI.Infrastructure.Configuration.DataAccess;
-using Energinet.DataHub.EDI.Infrastructure.InboxEvents;
 using Energinet.DataHub.EDI.IntegrationTests.Factories;
 using Energinet.DataHub.EDI.IntegrationTests.Fixtures;
 using Energinet.DataHub.EDI.MasterData.Interfaces;
+using Energinet.DataHub.EDI.Process.Infrastructure.Configuration.DataAccess;
+using Energinet.DataHub.EDI.Process.Infrastructure.InboxEvents;
 using Energinet.DataHub.Edi.Responses;
 using Google.Protobuf;
 using Google.Protobuf.WellKnownTypes;
@@ -88,7 +88,7 @@ public class WhenAggregatedTimeSeriesRequestAcceptedEventIsReceivedTests : TestB
 
     private void RegisterInboxEvent()
     {
-        var context = GetService<B2BContext>();
+        var context = GetService<ProcessContext>();
         context.ReceivedInboxEvents.Add(new ReceivedInboxEvent(
             _eventId,
             _eventType,
