+++ source/Process.Domain/Process.Domain.csproj
@@ -13,6 +13,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <PackageReference Include="Azure.Messaging.ServiceBus" Version="7.17.2"/>
     <PackageReference Include="NodaTime.Serialization.SystemTextJson" Version="1.1.2" />
   </ItemGroup>
