+++ source/Process.Infrastructure/InboxEvents/InboxEventsConfiguration.cs
@@ -17,7 +17,7 @@ using Energinet.DataHub.EDI.BuildingBlocks.Infrastructure.TimeEvents;
 using MediatR;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace Energinet.DataHub.EDI.Infrastructure.InboxEvents;
+namespace Energinet.DataHub.EDI.Process.Infrastructure.InboxEvents;
 
 public static class InboxEventsConfiguration
 {
