+++ source/Process.Infrastructure/InboxEvents/InboxEventsProcessor.cs
@@ -23,7 +23,7 @@ using Energinet.DataHub.EDI.Common.DateTime;
 using MediatR;
 using Microsoft.Extensions.Logging;
 
-namespace Energinet.DataHub.EDI.Infrastructure.InboxEvents;
+namespace Energinet.DataHub.EDI.Process.Infrastructure.InboxEvents;
 
 public class InboxEventsProcessor
 {
@@ -69,7 +69,9 @@ public class InboxEventsProcessor
             catch (Exception e)
             {
                 await MarkAsFailedAsync(inboxEvent, e, cancellationToken).ConfigureAwait(false);
+#pragma warning disable CA1727
                 _logger.LogError(e, "Failed to process inbox event. Id: {id}, EventType: {@Event}", inboxEvent.Id, inboxEvent.EventType);
+#pragma warning restore CA1727
             }
         }
     }
