+++ source/Process.Infrastructure/InboxEvents/ProcessInboxEventsOnTenSecondsHasPassed.cs
@@ -17,7 +17,7 @@ using System.Threading.Tasks;
 using Energinet.DataHub.EDI.BuildingBlocks.Infrastructure.TimeEvents;
 using MediatR;
 
-namespace Energinet.DataHub.EDI.Infrastructure.InboxEvents;
+namespace Energinet.DataHub.EDI.Process.Infrastructure.InboxEvents;
 
 public class ProcessInboxEventsOnTenSecondsHasPassed : INotificationHandler<TenSecondsHasHasPassed>
 {
